src/requirements.txt
cloudshell-networking-alcatel>=1.0,<1.1